make api restful
